Fix contentManagement name and route
